+++ Assets/Scripts/WeaponScripts/spear.cs
@@ -5,13 +5,13 @@ using UnityEngine;
 public class spear : MonoBehaviour
 {
 
-    public int atkDamage = 7;
-    public int characterDmg;
+    public float atkDamage = 7;
+    public float characterDmg;
     public float atkSpeed = 1.5f;
     float lastAttackTime;
     public float attackRange = 3f;
     public GameObject[] enemies;
-    public int totalatk;
+    public float totalatk;
     private bool attackMade;
 
     void Start()
